enhance context in retriever
